将Node相关的类独立放置于单独的文件中，避免循环导入问题。
