+++ auto_questions_v1_3/graph.py
@@ -4,54 +4,13 @@
 """命题推理使用的推理图
 """
 
-import element
-import config
 import proposition as prop
+import mynode
 import rule
-from enum import StrEnum
 import math
 from collections.abc import Sequence
 from typing import Optional
 from itertools import takewhile
-
-class NodeField(StrEnum):
-    """推理图中节点的字段
-    """
-    Condition = "condition" # 条件
-    Conclusion = "conclusion" # 结论
-    Rule = "rule" # 规则
-    Layer = "layer" # 层级
-    ConditionLayers = "condition_layers" # 条件的层级
-
-class Node(element.Element):
-    """推理图中的节点
-    """
-    def __init__(self, name = "", kind = "", **kwargs):
-        super().__init__(name, kind, **kwargs)
-        self[NodeField.ConditionLayers] = [math.inf] * len(self[NodeField.Condition]) # 条件的层级
-        self[NodeField.Layer] = math.inf # 默认层级为无穷大
-
-    def translate(self, lang, require = None, **kwargs):
-        # TODO: 推理图上节点的翻译方法，用于生成CoT
-        pass
-
-    def set_layer(self, curr_layer: int, curr_props: list[prop.Proposition]) -> Optional[prop.Proposition]:
-        """设置节点的层级
-
-        Args:
-            curr_layer (int): 当前层级
-            curr_props (list[prop.Proposition]): 当前的命题
-
-        Returns:
-            Optional[prop.Proposition]: 如果节点层数小于当前层数，返回节点的结论命题，否则返回None
-        """
-        conditions: list[prop.Proposition] = self[NodeField.Condition]
-        for i in takewhile(lambda x: self[NodeField.ConditionLayers][x] > curr_layer, range(len(conditions))):
-            if conditions[i].is_contained(curr_props):
-                self[NodeField.ConditionLayers][i] = min(self[NodeField.ConditionLayers][i], curr_layer)
-        self[NodeField.Layer] = max(self[NodeField.ConditionLayers])
-        # 如果节点层数小于当前层数，返回结论命题，否则返回None
-        return self[NodeField.Conclusion] if self[NodeField.Layer] < curr_layer else None
 
 class ReasoningGraph:
     """推理图，内含全面的推理结果，是程序的核心组件之一\n
@@ -70,14 +29,14 @@ class ReasoningGraph:
         self.init_props: list[prop.Proposition] = list(init_props) # 推理图中的初始命题
         self.reasoning_rules: list[rule.Rule] = list(rules) # 推理图中可用的推理规则
         self.knowledge_props: list[prop.Proposition] = list(knowledge_props) if knowledge_props is not None else []
-        self.nodes: list[Node] = [] # 推理图中的节点
+        self.nodes: list[mynode.Node] = [] # 推理图中的节点
         self.deepest_layer: int = -1 # 推理图中最深的层级
 
-    def add_nodes(self, nodes: Sequence[Node]):
+    def add_nodes(self, nodes: Sequence[mynode.Node]):
         """添加节点
 
         Args:
-            nodes (Sequence[Node]): 节点序列
+            nodes (Sequence[mynode.Node]): 节点序列
         """
         self.nodes.extend(nodes)
 
@@ -103,7 +62,7 @@ class ReasoningGraph:
         """
         conclusions: list[prop.Proposition] = []
         for node in self.nodes:
-            node_conclusion: prop.Proposition = node[NodeField.Conclusion]
+            node_conclusion: prop.Proposition = node[mynode.NodeField.Conclusion]
             if not node_conclusion.is_contained(conclusions):
                 conclusions.append(node_conclusion)
         return conclusions
@@ -122,7 +81,7 @@ class ReasoningGraph:
             curr_prop_list: list[prop.Proposition] = new_props + self.get_conclusions() + self.knowledge_props
         assert len(curr_prop_list) > 0, "没有命题可以推理"
         assert len(self.reasoning_rules) > 0, "没有推理规则"
-        curr_nodes: list[Node] = []
+        curr_nodes: list[mynode.Node] = []
         while True:
             reason_count += 1
             print(f"执行第{reason_count}次推理")
@@ -132,7 +91,7 @@ class ReasoningGraph:
                 print(f"执行规则{rule.name}成功")
                 curr_nodes.extend(rule_result)
             self.add_nodes(curr_nodes)
-            curr_conclusions: list[prop.Proposition] = [i[NodeField.Conclusion] for i in curr_nodes]
+            curr_conclusions: list[prop.Proposition] = [i[mynode.NodeField.Conclusion] for i in curr_nodes]
             if all([i.is_contained(curr_prop_list) for i in curr_conclusions]):
                 break
             # 将当前命题的FIRST_USED设置为True，表示已经被使用
@@ -148,15 +107,15 @@ class ReasoningGraph:
         """
         # 重置节点的层级
         for node in self.nodes:
-            node[NodeField.Layer] = math.inf
-            node[NodeField.ConditionLayers] = [math.inf] * len(node[NodeField.Condition])
+            node[mynode.NodeField.Layer] = math.inf
+            node[mynode.NodeField.ConditionLayers] = [math.inf] * len(node[mynode.NodeField.Condition])
         curr_layer_props: list[prop.Proposition] = chosen_props + self.knowledge_props
         next_layer_props: list[prop.Proposition] = []
         layer: int = 0
-        while any([i[NodeField.Layer] > layer for i in self.nodes]):
+        while any([i[mynode.NodeField.Layer] > layer for i in self.nodes]):
             layer += 1
             print(f"设置第{layer}层节点")
-            for node in takewhile(lambda x: x[NodeField.Layer] > layer, self.nodes):
+            for node in takewhile(lambda x: x[mynode.NodeField.Layer] > layer, self.nodes):
                 conclusion = node.set_layer(layer, curr_layer_props)
                 if conclusion and conclusion.is_contained(next_layer_props):
                     next_layer_props.append(conclusion)
@@ -175,8 +134,8 @@ class ReasoningGraph:
         """
         assert self.deepest_layer >= 0, "尚未进行二次推理"
         conclusion_list: list[prop.Proposition] = []
-        for node in takewhile(lambda x: x[NodeField.Layer] == self.deepest_layer, self.nodes):
-            node_conclusion: prop.Proposition = node[NodeField.Conclusion]
+        for node in takewhile(lambda x: x[mynode.NodeField.Layer] == self.deepest_layer, self.nodes):
+            node_conclusion: prop.Proposition = node[mynode.NodeField.Conclusion]
             if not node_conclusion.is_contained(conclusion_list):
                 conclusion_list.append(node_conclusion)
         return conclusion_list
